+++ mopidy/local/__init__.py
@@ -4,7 +4,7 @@ import logging
 import os
 
 import mopidy
-from mopidy import config, ext, models
+from mopidy import config, ext
 
 logger = logging.getLogger(__name__)
 
@@ -106,26 +106,12 @@ class Library(object):
         """
         Lookup the images for the given URIs.
 
-        The default implementation will simply call :meth:`lookup` and
-        try and use the album art for any tracks returned. Most local
-        libraries should replace this with something smarter or simply
-        return an empty dictionary.
+        The default implementation will simply return an empty dictionary.
 
         :param list uris: list of URIs to find images for
         :rtype: {uri: tuple of :class:`mopidy.models.Image`}
         """
-        result = {}
-        for uri in uris:
-            image_uris = set()
-            tracks = self.lookup(uri)
-            # local libraries may return single track
-            if isinstance(tracks, models.Track):
-                tracks = [tracks]
-            for track in tracks:
-                if track.album and track.album.images:
-                    image_uris.update(track.album.images)
-            result[uri] = [models.Image(uri=u) for u in image_uris]
-        return result
+        return {}
 
     def load(self):
         """