+++ mopidy/models/__init__.py
@@ -1,6 +1,5 @@
 from __future__ import absolute_import, unicode_literals
 
-from mopidy import compat
 from mopidy.models import fields
 from mopidy.models.immutable import ImmutableObject, ValidatedImmutableObject
 from mopidy.models.serialize import ModelJSONEncoder, model_json_decoder
@@ -144,12 +143,6 @@ class Album(ValidatedImmutableObject):
     :type date: string
     :param musicbrainz_id: MusicBrainz ID
     :type musicbrainz_id: string
-    :param images: album image URIs
-    :type images: list of strings
-
-    .. deprecated:: 1.2
-        The ``images`` field is deprecated.
-        Use :meth:`mopidy.core.LibraryController.get_images` instead.
     """
 
     #: The album URI. Read-only.
@@ -172,12 +165,6 @@ class Album(ValidatedImmutableObject):
 
     #: The MusicBrainz ID of the album. Read-only.
     musicbrainz_id = fields.Identifier()
-
-    #: The album image URIs. Read-only.
-    #:
-    #: .. deprecated:: 1.2
-    #:     Use :meth:`mopidy.core.LibraryController.get_images` instead.
-    images = fields.Collection(type=compat.string_types, container=frozenset)
 
 
 class Track(ValidatedImmutableObject):